Creacion clase pokemon y definiciones.h
