+++ src/pokemon.cpp
@@ -1,0 +1,6 @@
+#include "pokemon.h"
+
+Pokemon::Pokemon()
+{
+
+}
