+++ src/tipos.h
@@ -2,31 +2,10 @@
 #define TIPOS_H
 
 #include <string>
+#include "definiciones.h"
 
-typedef unsigned short int uint;
-using namespace std;
 
-enum Tipos : uint {
-    NINGUNO = 0,
-    NORMAL = 1,
-    FUEGO = 2,
-    LUCHA = 3,
-    AGUA = 4,
-    VOLADOR = 5,
-    PLANTA = 6,
-    ELECTRICO = 7,
-    VENENO = 8,
-    TIERRA = 9,
-    PSIQUICO = 10,
-    ROCA = 11,
-    HIELO = 12,
-    BICHO = 13,
-    DRAGON = 14,
-    FANTASMA = 15,
-    SINIESTO = 16,
-    ACERO = 17,
-    HADA = 18,
-};
+
 
 class Tipo{
     private:
